refactor(gateway): split service.go by concern
service.go held 1289 LOC spanning chat-completion routing, provider
catalog/readiness, usage/trace/retention queries, and shared
helpers. Pull the provider catalog and the telemetry-query methods
out into per-concern siblings; the chat flow keeps service.go.

  service.go:               1289 → 768 LOC (chat flow, types, streaming, validate)
  service_providers.go:     new, 460 LOC
  service_telemetry.go:     new, 84 LOC

What moves to service_providers.go

  ListModels, ProviderStatus, ProviderModelReadiness,
  ProviderHealthHistory (4 public methods) plus the
  provider-readiness helper cluster:
    providerCredentialReady, providerRoutingReady,
    providerRoutingBlockedReason, providerReadinessChecks,
    providerReadinessSummary, providerReadinessSummaryFromCheck,
    providerCredentialCheck, providerModelDiscoveryCheck,
    providerHealthCheck, providerReadinessHealthReason,
    providerHealthFailureMessage, providerHealthDegradedMessage,
    providerHealthPendingMessage, providerRoutingCheck,
    providerRoutingBlockedMessage, providerReadinessCheck,
    providerReadinessOperatorAction, pluralS.

What moves to service_telemetry.go

  UsageSummaryWithFilter, UsageEvents (2 governor-backed queries)
  ListTraces, Trace (2 profiler queries)
  RunRetention, ListRetentionRuns (2 retention manager calls)

What stays in service.go

  - Dependencies, ResilienceOptions, Service struct; NewService /
    Tracer construction.
  - Chat flow: HandleChat, recordRequestOutcome, buildExecutionPlan,
    executePlan, estimateUsage, withResolvedModel.
  - Streaming: StreamHandle methods, sseCapture type/methods,
    HandleChatStreamCapture, RouteForStream, fallbackFrom,
    normalizeResilienceOptions.
  - validate, captureRequestBody, captureResponseBody,
    redactSensitiveText.
  - All Result types (ChatResult, ModelsResult, ProviderStatusResult,
    etc.) stay in service.go to keep one home for the package's
    exported shape.

No public symbol moves to a different package; the receiver type
stays *Service everywhere; mux registrations + handler call sites
in internal/api/ are unchanged. Imports per file are minimal:
service_providers.go takes context/fmt/strings/time + catalog,
providers, pkg/types. service_telemetry.go takes context/fmt +
governor, retention.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
